LayoutTests/platform/blackberry/fast/filesystem/op-restricted-names-expected.txt

CONSOLE MESSAGE: line 394: ReferenceError: Can't find variable: webkitRequestFileSystem
FAIL: Timed out waiting for notifyDone to be called
